feat: add support for multi fills and fill format
Multi branch fill select: if the fill expression results in a node-set
(e.g. ["one", "two"]), and the attribute is part of a multi branch
(i.e. an ancestor has declared pt:multi), fill will select the node
of the node set corresponding to the attribute's multi branch index.
This allows for e.g. multiple class instantiation with different
child attribute values (both in a fetch context and not).

Fill formatting: if the attribute is populated and contains a "{}"
format token, substitute the token for the evaluated fill expression.
Only a single format token per attribute is currently supported.