fix: ignore pre/post whitespace in LIDs
Strip leading and trailing whitespace when interacting with LIDs. This
is a stop-gap measure; longer term we should e.g. define a function in
the label_tools module for retrieving attribute text (could be combined
with a node-type invariant unwrapper, e.g. as in ext.pt.context:_unpack)
to provide extension functions and others with a unified approach.

Related to #4, which could be broadened to cover internal whitespace
issues as there is some crossover with the final representation.