feat: allow fetched children to follow src order
Introduce `pt:reorder`, which if present on a fetched class will ensure
that the document order of its children corresponds as closely as
possible with that of the active source element. Children unique to
the template element will cling on to their originally proceeding
sibling, avoiding most PDS4 order related pitfalls.

Note that the only reason for introducing this property is to put out
the burning dumpster fire that is the interleaving of
{,Group_}Field_Binary children of dynamic cardinality in Table_Binary
classes, where document order unfortunately matters a lot.

This property should be revised at a later point when the dust settles,
as is not well aligned with the core PT mantra of WYSIWYG (more in #18).