refactor(ext-exm): rework LID handling
Attempt to clarify structure and option space for EXM LIDs while laying
the groundwork for future multi-instrument support on the PID front.

LIDs are now comparable based on their string representation, and
sorting is explicitly delegated to their PID component.

Currently only PanCamPID is defined; in future a ProductID base class
should be implemented with subclass delegation similar to that used for
ProcTools.DataProduct (here using the bundle ID as the discriminator).

Caching of ExoMarsLID instances is implemented using the new ext func
resources feature.