feat: add file handler i/f, ext funcs; integration
File handlers can now be associated with label structures and registered
to a template. Upon template export, it's file handlers will be prompted
to write their payloads to disk as well.

The resources feature is used to interface registered file handlers with
extension functions, which allows deferred population of key properties
such as size, offset and md5.

In additon to FileHandler, the DataObject interface is intended to allow
a range of specific PDS4 data object types to be realised, and
optionally dynamically instantiated during FileHandler creation.

BREAKING CHANGE: filename is now a required parameter of Template.export